Create perl_mutator.pl
Take input from a text file and converts it into a random string output file.

We'll use a regular expression to match URLs in the text file and replace them with random strings generated using Data::UUID.

In this updated code, we use a regular expression with the s/// operator to match URLs (starting with "http://" or "https://") in the input text file. We replace each matched URL with a random string generated using the Data::UUID module's _generate_random_string() subroutine. This ensures that each URL in the text file is replaced with a unique random string.