Create file_size_scanner.pl
This Perl script will prompt the user for the directory to scan and the maximum file size in MB. It will then scan the selected directory and display all files that are of the user-selected size and below. It will then ask the user to confirm if they wish to delete these files. If confirmed, the script will proceed to delete the files and display the number of files deleted.

Will *not* scan sub-directories.