Create ShannJensKL_EntropyCalc.pl
To create a program that calculates the Kullback-Leibler (KL) divergence and Jensen-Shannon (JS) divergence between the Shannon entropies of selected text files, we need to modify the script to calculate the entropies and then use those values to calculate the divergences.

Here's a complete Perl script that does just that:

This script will calculate the Shannon entropy for each selected file, and then it will calculate the KL divergence and JS divergence between all pairs of entropies for the selected files. The divergences are printed to the console. Remember to save this script with a .pl extension and run it using a Perl interpreter.